+++ Parylation/Parylation/Presentation/Scenes/Settings/SettingsBuilder.swift
@@ -24,16 +24,19 @@ final class SettingsBuilderImpl {
 // MARK: - SettingsBuilder implementation
 
 extension SettingsBuilderImpl: SettingsBuilder {
-    func build() -> UIViewController {
+    func build(listener: SettingsListener?) -> UIViewController {
         let view = SettingsView()
 
         let interactor = SettingsInteractorImpl(
-            authorizationService: context.authorizationService
+            credentialsValidatorService: CredentialsValidatorServiceImpl(),
+            authorizationService: context.authorizationService,
+            userService: context.userService
         )
         let alertBuilder = AlertBuilderImpl()
         let router = SettingsRouterImpl(
             view: view,
-            alertBuilder: alertBuilder
+            alertBuilder: alertBuilder,
+            listener: listener
         )
         let viewModel = SettingsViewModelImpl(
             interactor: interactor,